pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "metabase-fastmcp"
version = "0.1.0"
description = "A FastMCP server for Metabase integration"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Hugo Lau", email = "hugolau23@example.com"},
]
keywords = ["metabase", "mcp", "model-context-protocol", "ai", "claude", "cursor"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastmcp>=0.19.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.10.0",
    "isort>=5.13.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
]

[project.scripts]
metabase-mcp = "server:main"

[project.urls]
Homepage = "https://github.com/hyeongjun-dev/metabase-mcp"
Repository = "https://github.com/hyeongjun-dev/metabase-mcp"
Issues = "https://github.com/hyeongjun-dev/metabase-mcp/issues"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space" 